% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_assignConsensusPloidy.R
\name{assignConsensusPloidy}
\alias{assignConsensusPloidy}
\title{Assign Consensus Ploidy}
\usage{
assignConsensusPloidy(
  dt_input_fractions,
  input_chrom = NULL,
  max_ploidy_force = NULL
)
}
\arguments{
\item{dt_input_fractions}{Data.table generated by calWatsonFractions().}

\item{input_chrom}{Vector of chromosomes ("chr1, chr2...") supplied by user. Automatically extracts chromosomes from data set if not specified.}

\item{max_ploidy_force}{Maximum ploidy determined by number of cells within the sample. If not supplied by the user automatically determined from data set.}
}
\value{
Data.table with two new additional columns: "sil_k" is the optimal k determined by maximizing a silhouette score and "cons_ploidy" is the consensus ploidy value of each window inferred after kmeans clustering and k+1=n
}
\description{
This function finds the optimal k for the Watson-reads fraction distribution for
each segment determined using a sliding window by calWatsonFractions().
}
\examples{
# run ploidy assignment
K562_chr1_ploidy <- assignConsensusPloidy(data_K562_ploidy[,1:7], input_chrom = "chr1")
# plot results in karyogram style to inspect
fct_plot_karyogram(K562_chr1_ploidy)
}
