% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{cal_silhouette_score}
\alias{cal_silhouette_score}
\title{Calculate Silhouette Scores}
\usage{
cal_silhouette_score(cur_k, fraction_w_vector)
}
\arguments{
\item{cur_k}{Integer of k to be scored.}

\item{fraction_w_vector}{Vector of Watson-read fractions of a given segment to be clustered using kmeans.}
}
\value{
Silhouette score for the kmeans of a fraction_w vector using cur_k. If any NA is present in the input vector function returns NA.
}
\description{
The silhouette score for a given k of kmeans is calculated and returned.
}
\examples{
cal_silhouette_score(cur_k = 3, c(0,0.01,0.001,0.334,0.32,0.33,0.65,0.67,0.66,1.0,0.99,0.998))
cal_silhouette_score(cur_k = 4, c(0,0.01,0.001,0.334,0.32,0.33,0.65,0.67,0.66,1.0,0.99,0.998))
}
