% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_K562_ploidy}
\alias{data_K562_ploidy}
\title{Example Data: Results of ploidy analysis for K562 using PloidyAssignR}
\format{
A data frame with 473863 rows and 11 variables
\itemize{
  \item {cell} {Cell}
  \item {chrom} {Chromosome (chr1...chrX)}
  \item {start} {Start position of the sliding window}
  \item {end} {End position of the sliding window}
  \item {sample} {Sample name: K562}
  \item {total_count} {Total count number within the window}
  \item {fraction_w} {Proportion of W-oriented reads within the window}
  \item {sil_k} {optimized number of clusters k by maximizing a silhouette score}
  \item {cons_ploidy} {Ploidy state as determined from counting clusters n = k-1}
  \item {baseline_coverage} {Mean count number within the ROC divided by the specific consensus.}
  \item {norm_count} {Normalized count number for the window: total_count divided by the baseline_coverage }
}
}
\usage{
data(data_K562_ploidy)
}
\description{
A dataset containing the complete output after ploidy analysis using PloidyAssignR.
}
\keyword{datasets}
