% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_K562_ROC}
\alias{data_K562_ROC}
\title{Example Data: Regions of Confidence for K562}
\format{
A data frame with 6 rows and 3 variables
\itemize{
  \item {chrom} {Chromosome (chr1...chrX)}
  \item {start} {Start position of the ROC}
  \item {end} {End position of the ROC}
}
}
\usage{
data(data_K562_ROC)
}
\description{
A data set containing the start and end point of possible regions of confidence in
several chromosomes as determined by the authors of this package.
}
\keyword{datasets}
