% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_K562_strand_seq_count}
\alias{data_K562_strand_seq_count}
\title{Example Data: Strand-seq count data created by MosaiCatcher pipeline.}
\format{
A data frame with 2502414 rows and 8 variables
\itemize{
  \item {chrom} {Chromosome (chr1...chrX)}
  \item {start} {Start position of the bin}
  \item {end} {End position of the bin}
  \item {sample} {Sample name: K562}
  \item {cell} {Cell}
  \item {c} {number of Crick-oriented reads in this bin}
  \item {w} {number of Watson-oriented reads in this bin}
  \item {class} {Class attribute as determined by MosaiCatcher pipeline}
}
}
\usage{
data(data_K562_strand_seq_count)
}
\description{
A dataset containing the binned Strand-seq count data.
}
\keyword{datasets}
