% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{export_data}
\alias{export_data}
\title{Export Data Table}
\usage{
export_data(
  data_table_input,
  omitNA = FALSE,
  output_folder,
  add_file_name = "PloidyAssignR"
)
}
\arguments{
\item{data_table_input}{Any data.table that should be exported.}

\item{omitNA}{Default is FALSE. If set to TRUE rows containing NA will be removed before export.}

\item{output_folder}{Path to output folder.}

\item{add_file_name}{String containing the desired name of the exported file.}
}
\value{
Saves data.table to the desired output folder
}
\description{
Wrapper function to export table as .csv, uses fwrite which is faster than standard writing options in r
}
\examples{
\dontrun{
path_wd <- getwd()
export_data(data_K562_ploidy, output_folder = path_wd, add_file_name = "Internal_Example")
}
}
