% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{export_plot}
\alias{export_plot}
\title{export_plot}
\usage{
export_plot(
  input_plot,
  file_name = "PloidyAssignR",
  file_type = "pdf",
  file_width = 297,
  file_height = 210,
  input_dpi = 60
)
}
\arguments{
\item{input_plot}{ggplot that should be exported}

\item{file_name}{String containing the desired name of the exported plot}

\item{file_type}{pdf or png or any other type Cairo accepts}

\item{file_width}{Default is 297 mm}

\item{file_height}{Default is 210 mm}

\item{input_dpi}{Default is 60 dpi.}
}
\value{
Exports plot using Cairo either as pdf or png.
}
\description{
export_plot
}
\examples{
# generate plot
cars_plot <- ggplot2::ggplot(data= mtcars)+ggplot2::geom_point(ggplot2::aes(x=mpg, y=cyl))
# basic export
export_plot(cars_plot, file_name="cars_plot.png", file_type = "png")

}
