% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{extract_chromosomes}
\alias{extract_chromosomes}
\title{Chromosomes}
\usage{
extract_chromosomes(input_chrom)
}
\arguments{
\item{input_chrom}{a vector of chromosomes or a data.table that contains a column "chrom" with chromosomes.}
}
\value{
Vector of chromosomes ordered by ascending number 1 to X/Y
}
\description{
Returns a vector of ordered chromosomes.
}
\examples{
extract_chromosomes(c("chr1", "chr2"))
extract_chromosomes(data_K562_strand_seq_count[c("chr1", "chr2"), on = "chrom"])

}
