% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_counts_cellwise}
\alias{fct_plot_counts_cellwise}
\title{fct_plot_counts_cellwise}
\usage{
fct_plot_counts_cellwise(
  data_counts,
  data_SC_CN = NULL,
  selected_cells,
  input_chrom,
  lab_title = NULL
)
}
\arguments{
\item{data_counts}{Data.table containing Strand-seq count data}

\item{data_SC_CN}{Data.table containing SC normalized count data (SC copy number)}

\item{selected_cells}{Character-Vector of cells that should be plotted}

\item{input_chrom}{Specific chromosome that should be plotted}

\item{lab_title}{Title of plot}
}
\value{
ggplot bar graph using teal and orange to specifiy read orientation.
}
\description{
fct_plot_counts_cellwise
}
\examples{
plot_cell <- fct_plot_counts_cellwise(data_K562_strand_seq_count,
                            selected_cells = c("01|301", "01|304"),
                            input_chrom = "chr1")
}
