% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_distribution_patterns}
\alias{fct_plot_distribution_patterns}
\title{fct_plot_distribution_patterns}
\usage{
fct_plot_distribution_patterns(
  data_ploidy,
  input_chrom,
  point_size = 0.5,
  input_start = NULL,
  input_end = NULL,
  input_cell = NULL,
  reg_color = "black",
  cell_color = "green",
  cell_size = 0.5
)
}
\arguments{
\item{data_ploidy}{Data.table containing consensus ploidy values.}

\item{input_chrom}{Character vector containing the desired chromosome to be plotted.}

\item{point_size}{Default is 0.5. Size should be chosen depending on sample size.}

\item{input_start}{If supplied will only plot distribution patterns in specified region.}

\item{input_end}{If supplied will only plot distribution patterns in specified region.}

\item{input_cell}{If supplied highlights specified cell in distribution plot.}

\item{reg_color}{color of the scatter plot points}

\item{cell_color}{color of the highlighted input_cell}

\item{cell_size}{Point size of cell input. Equivalent to point size parameter of geom_point. T}
}
\value{
A ggplot of the distribution patterns.
}
\description{
fct_plot_distribution_patterns
}
\examples{
fct_plot_distribution_patterns(data_K562_ploidy, input_chrom = "chr1")
}
