% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_karyogram}
\alias{fct_plot_karyogram}
\title{fct_plot_karyogram}
\usage{
fct_plot_karyogram(data_ploidy, chrom_order = NULL)
}
\arguments{
\item{data_ploidy}{data.table containing consensus ploidy values.}

\item{chrom_order}{Optional: Desired order of chromosomes for plotting.}
}
\value{
Returns a ggplot bar graph using color-coding to show the consensus ploidy for chromosomes.
}
\description{
fct_plot_karyogram
}
\examples{
fct_plot_karyogram(data_K562_ploidy)
}
