% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_sc_heatmap}
\alias{fct_plot_sc_heatmap}
\title{fct_plot_sc_heatmap}
\usage{
fct_plot_sc_heatmap(
  data_ploidy,
  input_chrom,
  input_start = NULL,
  input_end = NULL
)
}
\arguments{
\item{data_ploidy}{Data.table containing the consensus ploidy values}

\item{input_chrom}{Desired chromosome that should be plotted.}

\item{input_start}{If supplied only plots SC ploidy values in the specified segment.}

\item{input_end}{If supplied only plots SC ploidy values in the specified segment.}
}
\value{
Returns a heatmap of the color-coded SC copy numbers.
}
\description{
fct_plot_sc_heatmap
}
\examples{
fct_plot_sc_heatmap(data_K562_ploidy, input_chrom = "chr1")
}
