% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{find_optimal_k}
\alias{find_optimal_k}
\title{Find optimal K}
\usage{
find_optimal_k(fraction_w_vector, max_ploidy = 4)
}
\arguments{
\item{fraction_w_vector}{Vector of Watson-reads fractions from one segments from all cells in sample set.}

\item{max_ploidy}{supplied by user and defaults to tetraploid.}
}
\value{
optimal k
}
\description{
Uses a silhouette score to determine optimal k in binomially distributed Watson-reads fraction data.
}
\examples{
find_optimal_k(c(0,0.01,0.001,0.334,0.32,0.33,0.65,0.67,0.66,1.0,0.99,0.998), max_ploidy=5)

}
