% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{format_Mb}
\alias{format_Mb}
\title{Format Mb}
\usage{
format_Mb(x)
}
\arguments{
\item{x}{any numeric that will be divided by 1e6}
}
\description{
readable output for x axis
}
\examples{
format_Mb(1000000)


}
