% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{input_prep}
\alias{input_prep}
\title{Input Preparation
Reads in Strand-seq library count file, checks formatting and prepares for analysis.}
\usage{
input_prep(input_data)
}
\arguments{
\item{input_data}{Path to or data.table of the Strand-seq counts file of one sample}
}
\value{
Data.table with necessary columns and sorted by setkey() to be more memory efficient.
}
\description{
Input Preparation
Reads in Strand-seq library count file, checks formatting and prepares for analysis.
}
\examples{
input_prep(data_K562_strand_seq_count)
}
