% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{ploidy_summary}
\alias{ploidy_summary}
\title{Ploidy Summary}
\usage{
ploidy_summary(data_ploidy)
}
\arguments{
\item{data_ploidy}{PloidyAssignR output containing consensus ploidy states.}
}
\value{
the mean ploidy state and the individual proportion of each ploidy state.
}
\description{
This function returns a data.table with the proportion of ploidy states.
}
\examples{
data_summary <- ploidy_summary(data_K562_ploidy)


}
