% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PloidyAssignR_run_all.R
\name{PloidyAssignR_run_all}
\alias{PloidyAssignR_run_all}
\title{PloidyAssignR run all analysis}
\usage{
PloidyAssignR_run_all(
  input_data,
  input_ROC = NULL,
  max_ploidy_force = NULL,
  input_chrom = NULL,
  input_window = 1e+07,
  input_step = 5e+06,
  export_path = NULL,
  export_file_name = "PloidyAssignR"
)
}
\arguments{
\item{input_data}{Path to MC counts file}

\item{input_ROC}{Path to file containing regions of confidence}

\item{max_ploidy_force}{Optional: Set max ploidy.}

\item{input_chrom}{Optional: Set chromosome selection for analysis.}

\item{input_window}{Optional: Set input window size. Default: 10000000}

\item{input_step}{Optional: Set input step size. Default: 5000000}

\item{export_path}{Optional export to given directory.}

\item{export_file_name}{Name that should be used to create exported files.}
}
\value{
data.table containing a column for Watson fractions, consensus ploidy, and SCnorm.
}
\description{
Wrapper function to run complete analysis including consensus and SC ploidy (if ROC file is available).
}
\examples{
analysis_results <- PloidyAssignR_run_all(
                        data_K562_strand_seq_count,
                        data_K562_ROC,
                        input_chrom = c("chr1", "chr18", "chr20")
                        )
# use internal plotting functions to visualize results: e.g. fct_plot_karyogram(analysis_results)
}
