% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scCoverageNorm.R
\name{scCoverageNorm}
\alias{scCoverageNorm}
\title{Single Cell Coverage Normalization}
\usage{
scCoverageNorm(dt_input_fractions, dt_input_ROC)
}
\arguments{
\item{dt_input_fractions}{data.table of Watson fractions}

\item{dt_input_ROC}{data.table of regions of confidence supplied by the user. Columns should be: chrom, start, end}
}
\value{
normalized read count that represents the single cell copy number
}
\description{
The baseline coverage of each cell library is determined from several
regions of confidence and from this the copy number of each segment is determined.
}
\examples{
analysis_results <- scCoverageNorm(data_K562_ploidy[,1:9], data_K562_ROC)
# use internal plotting methods to visualize results
fct_plot_sc_heatmap(analysis_results, input_chrom = "chr1")
}
