% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{template_input_ploidy_assignment}
\alias{template_input_ploidy_assignment}
\title{Example Structure of Input Data.}
\description{
Example Structure of Input Data.
}
\keyword{datasets}
