% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_assignConsensusPloidy.R
\name{assignConsensusPloidy}
\alias{assignConsensusPloidy}
\title{Assign Consensus Ploidy}
\usage{
assignConsensusPloidy(
  dt_input_fractions,
  input_chrom = NULL,
  max_ploidy_force = NULL
)
}
\arguments{
\item{dt_input_fractions}{Data.table from calWatsonFractions(), containing relative Watson.strand state frequencies fraction_w.}

\item{input_chrom}{(Optional) Vector that specifies chromosomes for analysis ("chr1, chr2..."). Defaults to all chromosomes if not set.}

\item{max_ploidy_force}{(Optional) Sets the maximum ploidy level, determined automatically from the number of cells if not specified.}
}
\value{
Returns the original data.able with added columns: "sil_k" is the optimal k determined by maximizing a silhouette score and "cons_ploidy" is the assigned consensus ploidy value for each window.
}
\description{
assignConsensusPloidy is a critical function in the PloidyAssignR algorithm, responsible for determining cell ploidy using k-means clustering.
It operates on Watson-reads fraction distributions from calWatsonFractions().
The function's key feature is the use of the silhouette score to determine the optimal number of clusters (k).
By determining the number of clusters, the tool detects the underlying consensus ploidy state.
}
\examples{
# Ploidy assignment for chromosome 1 in K562 cells
K562_chr1_ploidy <- assignConsensusPloidy(data_K562_ploidy[,1:7], input_chrom = "chr1")

# Karyogram visualization of the results
fct_plot_karyogram(K562_chr1_ploidy)
}
