% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calWatsonFractions.R
\name{calWatsonFractions}
\alias{calWatsonFractions}
\title{Calculate Watson Fractions using a Sliding Window}
\usage{
calWatsonFractions(
  dt_input_count,
  input_chrom = NULL,
  input_window = 1e+07,
  input_step = 5e+06,
  na_rm = TRUE
)
}
\arguments{
\item{dt_input_count}{The strand-seq count data.table that has been prepared by input_prep().}

\item{input_chrom}{= NULL ; if not supplied by user, all chromosomes in the data set supplied will be analyzed.}

\item{input_window}{= 10000000}

\item{input_step}{= 5000000}

\item{na_rm}{= TRUE ; NA throw errors and are removed from dataset.}
}
\value{
A data.table containing the columns "chrom", "start", "end", "sample", "cell", "total_count", "fraction_w".
}
\description{
Strand-seq count data is used to calculate the watson fractions for a sliding window.
}
\examples{
# calculate watson fractions
K562_chr1_windows <- calWatsonFractions(data_K562_strand_seq_count, input_chrom = "chr1")
# use internal plotting to visualize patterns fct_plot_distribution_patterns()
# then continue analysis with assignConsensusPloidy()...
}
