% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_distribution_patterns}
\alias{fct_plot_distribution_patterns}
\title{fct_plot_distribution_patterns}
\usage{
fct_plot_distribution_patterns(
  data_ploidy,
  input_chrom,
  point_size = 0.5,
  input_start = NULL,
  input_end = NULL,
  input_cell = NULL,
  reg_color = "black",
  cell_color = "green",
  cell_size = 1.5
)
}
\arguments{
\item{data_ploidy}{Data.table containing consensus ploidy values.}

\item{input_chrom}{Character vector containing the desired chromosome to be plotted.}

\item{point_size}{(Default: 0.5) Size of the scatter plot points. Adjust based on sample size.}

\item{input_start}{(Optional) Start position for plotting a specific genomic region.}

\item{input_end}{(Optional) End position for plotting a specific genomic region.}

\item{input_cell}{(Optional) Cell identifier to highlight in the distribution plot.}

\item{reg_color}{(Optional) Color of the scatter plot points.}

\item{cell_color}{(Optional) Color used to highlight the specified cell.}

\item{cell_size}{(Optional) Size of the scatter plot points of the specified cell.}
}
\value{
Produces a ggplot object displaying the distribution patterns of consensus ploidy.
This plot aids in visualizing the clustering of the relative W-strand state frequencies along the specified chromosome.
}
\description{
This function generates scatter plots that help with visualizing the distribution patterns of the calculated relative W-strand state frequencies fraction_w.
The y-axis shows the relative W-strand state frequency of each cell against the specific genomic position on the x-axis.
This plot allows comparative analysis of the ploidy-based distribution patterns across multiple cells and chromosomes.
}
\examples{
# Plot distribution patterns for chromosome 1
fct_plot_distribution_patterns(data_K562_ploidy, input_chrom = "chr1")
}
