% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fct_plot_sc_heatmap}
\alias{fct_plot_sc_heatmap}
\title{fct_plot_sc_heatmap}
\usage{
fct_plot_sc_heatmap(
  data_ploidy,
  input_chrom,
  input_start = NULL,
  input_end = NULL
)
}
\arguments{
\item{data_ploidy}{Data.table containing normalized count data for single cells.}

\item{input_chrom}{Specifies the chromosome for which the heatmap is to be generated.}

\item{input_start}{(Optional) Start position of the genomic region for a more focused heatmap.}

\item{input_end}{(Optional) End position of the genomic region for a more focused heatmap.}
}
\value{
Returns a heatmap visualization created using the pheatmap package. The heatmap displays the variation in normalized counts across cells,
with each cell represented as a row and genomic positions (formatted in Mb) as columns. The copy number is color-coded as displayed by a legend.
}
\description{
fct_plot_sc_heatmap is a function in the PloidyAssignR toolkit designed to visualize normalized count data
as a heatmap, specifically for single-cell analysis. The function focuses on displaying the variation in normalized counts
across cells within a specified genomic region. y representing normalized counts in a heatmap format, the function aids in the identification of
patterns and anomalies in chromosomal segments across different cells.
}
\examples{
# Create a heatmap for chromosome 1 of a single-cell data set
heatmap_plot <- fct_plot_sc_heatmap(data_K562_ploidy, input_chrom = "chr1")
}
