% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{input_prep}
\alias{input_prep}
\title{Input Preparation
The input_prep function performs the initial processing of Strand-seq library count data to prepare for ploidy analysis by PloidyAssignR.
It focuses on reading the input data (either from a path or directly as a data.table), verifying its format, and preparing it for subsequent analysis.
This ensures that the input data adheres to the required structure and format.}
\usage{
input_prep(input_data)
}
\arguments{
\item{input_data}{A file path or a data.table containing the Strand-seq counts for one sample.}
}
\value{
Data.table with necessary columns and sorted by setkey() to be more memory efficient.
}
\description{
Input Preparation
The input_prep function performs the initial processing of Strand-seq library count data to prepare for ploidy analysis by PloidyAssignR.
It focuses on reading the input data (either from a path or directly as a data.table), verifying its format, and preparing it for subsequent analysis.
This ensures that the input data adheres to the required structure and format.
}
\examples{
# Prepare input Strand-seq count data for PloidyAssignR analysis
prepared_data <- input_prep(data_K562_strand_seq_count)
}
