% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PloidyAssignR_run_all.R
\name{PloidyAssignR_run_all}
\alias{PloidyAssignR_run_all}
\title{PloidyAssignR run all analysis}
\usage{
PloidyAssignR_run_all(
  input_data,
  input_ROC = NULL,
  max_ploidy_force = NULL,
  input_chrom = NULL,
  input_window = 1e+07,
  input_step = 5e+06,
  export_path = NULL,
  export_file_name = "PloidyAssignR"
)
}
\arguments{
\item{input_data}{Path to Strand-seq count data file, as generated by MosaiCatcher.}

\item{input_ROC}{(Optional) Path to the file containing regions of confidence.}

\item{max_ploidy_force}{(Optional) Specifies the maximum ploidy level.}

\item{input_chrom}{(Optional) Specifies chromosomes for analysis.}

\item{input_window}{(Optional) The size of the sliding window for the calculation of the relative W-strand state frequency.}

\item{input_step}{(Optional) The step size for the sliding window.}

\item{export_path}{(Optional) Directory path for exporting results.}

\item{export_file_name}{(Optional) Name of the exported files.}
}
\value{
This wrapper function returns the final data.table containing columns with the relative W-strand state frequencies fraction_w, consensus ploidy, and the results of single-cell copy number assignment (SCnorm).
}
\description{
Execution of PloidyAssignR - Wrapper Function
}
\examples{
# Run complete PloidyAssignR analysis for specific chromosomes in K562 cells
analysis_results <- PloidyAssignR_run_all(
   data_K562_strand_seq_count,
   data_K562_ROC,
   input_chrom = c("chr1", "chr18", "chr20")
)

# Visualize the results using internal plotting functions
fct_plot_karyogram(analysis_results)
}
